Update Turkish abbreviation list
Processed TDK's list to extract abbr which uses periods (removed 2 instances already existing here). These are added with 1000 in the first column.
https://tdk.gov.tr/wp-content/uploads/2019/01/Kısaltmalar_Dizini.pdf

Also added single capital letters with period, which are used to shorten names.  These are added with 1000 in the first column. (we already had T. on the list)